Add first unit tests for UserDb2
